requirements.txt
# DAQS Clean Project Requirements
# Core dependencies for the streamlined DAQS framework

# Web frameworks
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# AI/ML
google-genai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0

# Data processing
pydantic>=2.0.0
pyyaml>=6.0.0
dataclasses-json>=0.5.0
jinja2>=3.0.0

# Visualization
graphviz>=0.20.0

# Logging
coloredlogs>=15.0.0

# Utilities
pathlib>=1.0.0
typing>=3.7.0